--- src/config/rest/index.ts
@@ -1,3 +1,0 @@
-export {RestConfig} from './rest.config.js';
-export {configRestSchema} from './rest.schema.js';
-export {RestAppConfig, RestSchema} from './rest.types.js';
