--- src/lib/logger/index.ts
@@ -1,2 +1,0 @@
-export {Logger} from './logger.interface.js';
-export {PinoLogger} from './pino.logger.js';
