+++ src/modules/config/index.ts
@@ -1,0 +1,3 @@
+export {RestConfig} from './rest-config.config.js';
+export {configRestSchema} from './rest-config.schema.js';
+export {RestAppConfig, RestSchema} from './types/rest-config.types.js';
