+++ src/modules/config/types/config.interface.ts
