+++ src/modules/logger/index.ts
@@ -1,0 +1,2 @@
+export {PinoLogger} from './pino.logger.js';
+export {Logger} from './types/logger.interface.js';
