+++ src/modules/logger/pino.logger.ts
@@ -1,7 +1,7 @@
 import { injectable } from 'inversify';
 import { Logger as PinoInstance, pino, } from 'pino';
 
-import { Logger } from './logger.interface.js';
+import { Logger } from './types/logger.interface.js';
 
 @injectable()
 export class PinoLogger implements Logger {
