+++ src/modules/rest/index.ts
@@ -1,0 +1,1 @@
+export {RestApp} from './app.rest.js';
