--- src/rest/index.ts
@@ -1,1 +1,0 @@
-export {RestApp} from './rest.app.js';
